Create README
The template contains preliminary pages as dictated by the guidelines. It also contains all the five chapters and what a student is expected to write in each of the five chapters. There is also a back matter section consisting of bibliography or references, glossary and appendix.  This template is developed using R Markdown in RStudio.  Since students are expected to present their thesis in Word document format, a .docx style file is included to enable you present your thesis in the required style. In addition, a Citation Style Language (CSL) is included to automate the process of writing references or bibliography.